+++ lib/blocks/text_fragment.dart
@@ -1,0 +1,68 @@
+import 'package:collection/collection.dart';
+
+class TextFragment {
+  /// The main data object associated with the line.
+  Object? data;
+
+  /// Optional attributes associated with the line data.
+  Map<String, dynamic>? attributes;
+
+  /// Constructs a [Line] instance with optional initial [data] and [attributes].
+  TextFragment({
+    this.data,
+    this.attributes,
+  });
+
+  /// Sets the data object of the line to [data].
+  void setData(Object? data) {
+    this.data = data;
+  }
+
+  /// Sets the attributes of the line to [attrs].
+  ///
+  /// If [attrs] is `null`, no changes are made to the current attributes.
+  void setAttributes(Map<String, dynamic>? attrs) {
+    if (attrs == null) return;
+    attributes = attrs;
+  }
+
+  /// Merges additional [attrs] into the current attributes.
+  ///
+  /// If [attributes] is `null`, creates a new map and adds [attrs] to it.
+  void mergeAttributes(Map<String, dynamic> attrs) {
+    attributes?.addAll(attrs);
+  }
+
+  /// Creates a deep copy of the current [Line] instance.
+  TextFragment get clone => TextFragment(data: data, attributes: attributes);
+
+  /// Clears the data and attributes of the line, setting them to `null`.
+  void cleanLine() {
+    data = null;
+    attributes = null;
+  }
+
+  @override
+  String toString() {
+    data ??= null;
+    attributes ??= null;
+    return 'TextFragment: "${data is String ? '$data'.replaceAll('\n', '\\n') : data}"${attributes == null ? '' : ', attributes: $attributes'}';
+  }
+
+  @override
+  bool operator ==(covariant TextFragment other) {
+    if (identical(this, other)) return true;
+    return (data is Map && other.data is Map
+            ? _equality.equals(data as Map, other.data as Map)
+            : data == other.data) &&
+        _equality.equals(
+          attributes,
+          other.attributes,
+        );
+  }
+
+  @override
+  int get hashCode => Object.hash(data, attributes);
+}
+
+const MapEquality _equality = MapEquality();