+++ lib/easy_parser/easy_parser.dart
@@ -3,20 +3,42 @@ import 'package:flutter_quill_delta_easy_parser/extensions/extensions.dart';
 import 'package:flutter_quill_delta_easy_parser/extensions/helpers/map_helper.dart';
 import 'package:flutter_quill_delta_easy_parser/flutter_quill_delta_easy_parser.dart';
 
-// Note: If you want to see how can look the [Document]s, take a look the test folder
 /// Represents a parser that converts the Quill Delta operations into a structured document format.
 class RichTextParser {
+  RichTextParser({
+    this.mergerBuilder = const CommonMergerBuilder(),
+  });
+
+  
+  /// This is the encharge to merge some paragraphs when they contains the same block attributes
+  /// or when contains same types.
+  ///
+  /// Default implementations:
+  ///
+  ///  1. [NoMergeBuilder]: don't do nothing
+  ///  2. [CommonMergerBuilder] (default merge behavior): check if the [Paragraph] can be merged. It's focused on merge general [Paragraph] (even if them are pure inline types) 
+  ///  3. [BlockMergerBuilder]: check just if the [Paragraph]s with block-attributes can be merge into a same one. 
+  ///
+  /// Example:
+  ///
+  /// ```dart
+  /// // to ignore merging behavior 
+  /// final parser1 = RichTextParser(mergerBuilder: NoMergeBuilder())
+  /// // to merge [Paragraph]s if them can do it 
+  /// final parser2 = RichTextParser(mergerBuilder: CommonMergerBuilder())
+  /// // to only merge blocks
+  /// final parser3 = RichTextParser(mergerBuilder: BlockMergerBuilder())
+  /// ```
+  final MergerBuilder mergerBuilder;
   final Document _document = Document(
     paragraphs: [],
   );
-  bool _isNumberedListActive = false;
 
   /// Parses a Quill Delta into a structured document.
   ///
   /// * [returnNoSealedCopies] indicates if will need to return a deep copy of the elements to avoid return a [Paragraph]s that cannot add more elements
   /// * [ignoreAllNewLines] indicates that all the new lines with no block-level target to apply will be ignored
   ///
-  /// Returns the parsed [Document] instance, or null if the delta is empty.
   Document? parseDelta(
     fq.Delta delta, {
     bool returnNoSealedCopies = false,
@@ -24,25 +46,35 @@ class RichTextParser {
   }) {
     if (delta.isEmpty) return null;
     _document.clean();
-    _isNumberedListActive = false;
-    final List<fq.Operation> denormalizedOperations =
-        delta.denormalize().operations;
+    final List<fq.Operation> denormalizedOperations = delta.denormalize().operations;
     bool ignoreNewLine = true;
     bool hasNextOp = true;
-    int? ignoreNewLineAtIndex = 0;
+    int? ignoreNewLineAtIndex;
+    // sometimes, we can find only new lines at the start of the Delta, then to avoid remove them, we
+    // will need to add a verification
+    bool startParagraphNewLineChecking = false;
+    Map<String, dynamic>? lastBlockAttributesKnown;
     for (int index = 0; index < denormalizedOperations.length; index++) {
       final fq.Operation operation = denormalizedOperations.elementAt(index);
-      final fq.Operation? nextOp =
-          denormalizedOperations.elementAtOrNull(index + 1);
+      final fq.Operation? nextOp = denormalizedOperations.elementAtOrNull(index + 1);
       // a basic check to avoid process retain or delete operations
       if (!operation.isInsert || nextOp != null && !nextOp.isInsert) {
-        final type =
-            nextOp != null && nextOp.isInsert ? nextOp.key : operation.key;
+        final type = nextOp != null && nextOp.isInsert ? nextOp.key : operation.key;
         throw StateError(
           'Operation at ${nextOp?.isInsert == false ? index + 1 : index} '
           'is "$type" type and parseDelta() only accepts "insert" types',
         );
       }
+      if (!startParagraphNewLineChecking) {
+        startParagraphNewLineChecking = operation.data != '\n';
+      }
+
+      if(operation.data == '\n' && !startParagraphNewLineChecking) {
+        _document.insert(Paragraph.newLine());
+        continue;
+      }
+
+
       ignoreNewLine = operation.data == '\n' && operation.attributes == null;
       hasNextOp = nextOp != null;
       // check if we should ignore the first new line after this operation
@@ -52,10 +84,12 @@ class RichTextParser {
       // "Paragraph 1, \n, Paragraph 2" => should be parsed to be => "Paragraph 1, Paragraph 2"
       //
       // "Paragraph 1, \n, \n, \n, Paragraph 2" => should be parsed to be => "Paragraph 1, \n, \n, Paragraph 2"
-      if (operation.data != '\n' &&
-          nextOp?.data == '\n' &&
-          nextOp?.attributes == null) {
+      if (operation.data != '\n' && nextOp?.data == '\n' && nextOp?.attributes == null) {
         ignoreNewLineAtIndex = index + 1;
+      }
+      if (operation.data == '\n' && operation.attributes == null && nextOp != null && nextOp.data != '\n') {
+        ignoreNewLine = true;
+        ignoreNewLineAtIndex = null;
       }
 
       // Verify first if the current if a line with just a new line before the last one that is the definitive block attribute
@@ -66,9 +100,7 @@ class RichTextParser {
       // When verify that the next operation has not the same attrs, then will ignore that new line since
       // that one is the definitive (it was the unique insert with the block attribute, but
       // denormalizer makes this to do more easy store on it)
-      if (nextOp != null &&
-              mapEquality(operation.attributes, nextOp.attributes) ||
-          !hasNextOp) {
+      if (nextOp != null && mapEquality(operation.attributes, nextOp.attributes) || !hasNextOp) {
         ignoreNewLine = false;
       }
 
@@ -81,12 +113,42 @@ class RichTextParser {
         ignoreNewLine = true;
       }
 
+      // current op is last
+      if (nextOp == null) {
+        _insertNewLine(
+          operation,
+          nextOp,
+          false,
+          ignoreAllNewLines,
+          false,
+          false,
+        );
+        break;
+      }
+
       _parseOperation(
         operation,
+        nextOp,
         ignoreNewLine,
         hasNextOp,
         ignoreAllNewLines,
+        nextOp.data == '\n' && nextOp.attributes != null,
       );
+
+      final Paragraph? lastPr = _document.getLast();
+      if (lastPr == null) {
+        lastBlockAttributesKnown = null;
+        continue;
+      }
+      if (lastBlockAttributesKnown != null &&
+          operation.data == '\n' &&
+          !mapEquality(lastBlockAttributesKnown, lastPr.blockAttributes)) {
+        lastPr.seal();
+        _document.updateLast(lastPr);
+        _startNewParagraph();
+      } else if (operation.data == '\n') {
+        lastBlockAttributesKnown = operation.attributes == null ? null : {...?operation.attributes};
+      }
     }
     // remove last if needed
     if (_document.paragraphs.isNotEmpty && _document.getLast()!.lines.isEmpty) {
@@ -96,6 +158,14 @@ class RichTextParser {
       if (_document.getLastSafe().isNewLine) {
         _document.paragraphs.removeLast();
       }
+    }
+    if (mergerBuilder.enabled) {
+      final List<Paragraph> paragraphs = <Paragraph>[..._document.paragraphs];
+      _document.clean();
+      final Iterable<Paragraph> newParagraphs = mergerBuilder.buildAccumulation(
+        paragraphs,
+      );
+      _document.paragraphs.addAll(newParagraphs);
     }
     if (returnNoSealedCopies) {
       return Document(
@@ -111,19 +181,23 @@ class RichTextParser {
 
   /// Internal method to parse a single Quill operation.
   void _parseOperation(
-    fq.Operation operation, [
+    fq.Operation operation,
+    fq.Operation? nextOperation, [
     bool ignoreNewLine = true,
     bool hasNextOp = false,
     bool ignoreAllNewLines = false,
+    bool nextIsBlockLevelAttributes = false,
   ]) {
     if (operation.data is Map) {
       _insertEmbed(operation, hasNextOp);
     } else if (operation.data == '\n') {
       _insertNewLine(
         operation,
+        nextOperation,
         ignoreNewLine,
         ignoreAllNewLines,
         hasNextOp,
+        nextIsBlockLevelAttributes,
       );
     } else {
       _insertText(operation, hasNextOp);
@@ -138,88 +212,131 @@ class RichTextParser {
     final Paragraph? lastPr = _document.getLast();
     if (lastPr != null && lastPr.lines.isEmpty) {
       lastPr
-        ..insert(Line(
-          data: operation.data,
+        ..insert(Line.fromData(
+          data: operation.data!,
           attributes: operation.attributes,
         ))
         ..seal();
       _document.updateParagraph(lastPr);
-      _isNumberedListActive = false;
       return;
     }
-    _document.insert(Paragraph.fromEmbed(operation)..seal());
-    _isNumberedListActive = false;
+    if (lastPr != null && lastPr.lines.isNotEmpty && lastPr.lines.first.isEmpty) {
+      lastPr
+        ..updateLine(
+            0,
+            Line.fromData(
+              data: operation.data!,
+              attributes: operation.attributes,
+            ))
+        ..seal();
+      _document.updateParagraph(lastPr);
+      return;
+    }
+    _document.insert(Paragraph.fromEmbed(operation));
   }
 
   /// Handles the insertion of a new line in the document.
   void _insertNewLine(
     fq.Operation operation,
+    fq.Operation? nextOperation,
     bool ignoreNewLine,
     bool ignoreAllNewLines,
     bool hasNextOp,
+    bool nextIsBlockLevelAttributes,
   ) {
     if (operation.attributes != null) {
-      Paragraph paragraph = _document.getLastSafe();
-      if (paragraph.isEmbed) {
-        _document
-            .updateParagraph(paragraph..blockAttributes = operation.attributes);
+      Paragraph? paragraph = _document.getLast();
+      if (paragraph == null) {
+        _document.insert(Paragraph.newLine());
+        return;
+      }
+      if (paragraph.isSealed) {
+        _startNewParagraph();
+      }
+      if (paragraph.isEmbed || paragraph.isTextInsert) {
+        if (paragraph.isTextInsert) {
+          paragraph.setType(ParagraphType.block);
+        }
+        if (nextIsBlockLevelAttributes) {
+          paragraph.seal();
+        }
+        _document.updateParagraph(paragraph..blockAttributes = operation.attributes);
         return;
       }
       // if the last added paragraph is already a block or line-break element
       // we need to add it as another element
       bool needIgnoreUpdate = false;
-      if (paragraph.isBlock || paragraph.isNewLine) {
-        if (!paragraph.isSealed) {
-          _document.updateParagraph(
-            paragraph..seal(),
-          );
-        }
+      if (paragraph.isSealed) {
         needIgnoreUpdate = ignoreAllNewLines;
-        paragraph = Paragraph(
-          lines: [
-            Line(data: operation.data),
-          ],
-          type: ParagraphType.lineBreak,
-        )..seal();
-      }
-      paragraph.blockAttributes = operation.attributes;
-      if (!_document.getLastSafe().isEmbed) {
-        paragraph.setType(ParagraphType.block);
+        paragraph = Paragraph.newLine();
       }
       if (!needIgnoreUpdate) {
-        _document.updateParagraph(paragraph);
-      }
-      if (operation.attributes?['list'] == 'ordered') {
-        if (!_isNumberedListActive) {
-          _isNumberedListActive = true;
+        paragraph.blockAttributes = operation.attributes;
+        if (!_document.getLastSafe().isEmbed) {
+          if (!paragraph.isNewLine) {
+            paragraph.setType(ParagraphType.block);
+          }
         }
-      } else {
-        _isNumberedListActive = false;
+        if (_document.getLastSafe().isEmpty) {
+          _document.updateLastSafe(paragraph);
+          return;
+        }
+        _document.updateParagraph(paragraph);
       }
       return;
     } else {
+      Paragraph? paragraph = _document.getLast();
+      paragraph ??= Paragraph.base();
       // if we have a paragraph that is currently empty, we use it instead create a new one
-      if (_document.getLast() != null && _document.getLast()!.lines.isEmpty) {
-        final Paragraph paragraph = _document.getLastSafe();
+      if (paragraph.isEmpty) {
         paragraph
-          ..insert(Line(data: '\n'))
+          ..insert(Line.newLine())
           ..setType(ParagraphType.lineBreak)
           ..seal();
-        _document.updateParagraph(paragraph);
-        if (hasNextOp) {
+        _document.getLast()?.seal();
+        _document.updateParagraphSafe(paragraph);
+        _startNewParagraph();
+        return;
+      }
+      if (!ignoreNewLine || !hasNextOp) {
+        _document.getLast()?.seal();
+        bool needNewParagraph = false;
+        if (paragraph.last != null && paragraph.last!.isEmpty) {
+          paragraph.removeLastLine();
+          paragraph.last?.seal();
+          needNewParagraph = true;
+        }
+        if (paragraph.isEmbed || paragraph.isNewLine || paragraph.isSealed || !hasNextOp) {
+          final Paragraph newLine = Paragraph.newLine();
+          _document.insert(newLine);
+          return;
+        }
+        final Paragraph newLine = Paragraph.newLine();
+        _document.insert(newLine);
+        if (needNewParagraph) {
           _startNewParagraph();
         }
         return;
       }
-      if (!ignoreNewLine || !hasNextOp) {
-        final Paragraph paragraph = Paragraph(
-          lines: [Line(data: '\n')],
-          type: ParagraphType.lineBreak,
-        );
-        _document.insert(paragraph..seal());
+      if (nextIsBlockLevelAttributes) {
+        _document.updateParagraphSafe(paragraph..seal());
       }
-      if (hasNextOp) {
-        _startNewParagraph();
+      // next is only a new line
+      if (!nextIsBlockLevelAttributes && nextOperation?.data == '\n' && !paragraph.isTextInsert) {
+        paragraph.seal();
+        _document.updateParagraph(paragraph);
+        return;
+      }
+      if (ignoreNewLine && hasNextOp && !nextIsBlockLevelAttributes) {
+        paragraph.last?.seal();
+        _document.updateParagraphSafe(paragraph);
+        if (nextOperation?.data == '\n' && nextOperation?.attributes == null) {
+          if (_document.getLast()?.last?.isEmpty ?? false) {
+            _document.getLast()?.removeLastLine();
+          }
+          _startNewParagraph();
+        }
+        return;
       }
     }
   }
@@ -227,13 +344,17 @@ class RichTextParser {
   /// Inserts text into the document.
   void _insertText(fq.Operation operation, bool hasNextOp) {
     Paragraph? paragraph = _document.getLast();
-    if (paragraph == null ||
-        paragraph.type != ParagraphType.inline ||
-        paragraph.isSealed) {
+    if (paragraph == null || paragraph.isSealed) {
       paragraph = Paragraph.base();
+      _document.insert(paragraph);
     }
-    paragraph.insert(
-      Line(
+    if (paragraph.isEmpty || (paragraph.last!.isSealed && paragraph.last!.isNotEmpty)) {
+      paragraph.insert(Line(
+        fragments: [],
+      ));
+    }
+    paragraph.insertTextFragment(
+      TextFragment(
         data: operation.data,
         attributes: operation.attributes,
       ),