+++ lib/plugins.dart
@@ -1,0 +1,3 @@
+library;
+
+export 'plugins/markdown/convert_to_markdown.dart';