+++ lib/plugins/markdown/convert_to_markdown.dart
@@ -1,0 +1,9 @@
+import 'package:flutter_quill_delta_easy_parser/flutter_quill_delta_easy_parser.dart';
+
+class MarkdownParser {
+  String parse(Document document) {
+    if(document.paragraphs.isEmpty) return '';
+    final String result = '';
+    return result; 
+  }
+}