+++ CHANGELOG.md
@@ -1,3 +1,45 @@
+## 1.1.0
+
+### Breaking changes
+
+There has been some confusion about why `Paragraph` contains a List of `Line` objects that are really just a portion of text. As mentioned, a `Line` should represent different lines of a `Paragraph`.
+
+It is because of these drawbacks that the `Line` class has been redesigned so that instead of containing a portion of the text within the `Paragraph`, it contains a new class called `TextFragment`.
+
+#### `TextFragment`
+
+Summarize: it's practically the same as `Line` was before, but with another name.
+
+#### `Line` Redesign
+
+`Line` is now able to behave as what it is, a line completely separate from its siblings.
+
+```dart
+class Line {
+  final List<TextFragment> _fragments;
+  final String id;
+  bool _sealed;
+
+  Line({
+    required List<TextFragment> fragments,
+  });
+
+  // General methods
+  List<TextFragment> get fragments;
+  void removeFragment(TextFragment fragment);
+  void addFragment(TextFragment fragment);
+  void updateFragment(int index, TextFragment fragment);
+}
+```
+
+* Feat: added support for create accumulator algorithms. 
+* Feat: added plugin to parser `Document` to `Markdown`. 
+* Fix: issues where the first new lines of the `Delta` are being remove unnecessarily.
+* Fix: issues where the `Paragraph`s with only a new-line is considered a `ParagraphType.block`.  
+* Fix: `toPrettyString()` bad string return.
+* Chore: added `toPrettyString()` for `Paragraph` class.
+* Chore(test): improve expect messages to be more readable.
+
 ## 1.0.6
 
 * Fix: `toPrettyString()` from Document that is not working as expected.